Moved client to separate file, and added client cjs module build, some cleanup and docs update.
